## Motivation and Context
Using llvm7, improve performance of X86 version.

## Description
almost same as static link support with llvm7(FPGA)
modified
　dlk/python/dlk/templates/Makefile.tpl

## How has this been tested?
run on X86 and FPGA(just to be sure).
run blueoil_test.sh

## Types of changes
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature / Optimization (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)