@lm-konda 2019-01-08
Performance change
before
```
konda@leapmind-server-08:~/work/blueoil/lmnet/saved/konda_20190108151918/export/save.ckpt-957/128x128/output/models/lib$ ./lm_x86.elf input.npy output.npy 
-------------------------------------------------------------
Comparison: Default network test  succeeded!!!
-------------------------------------------------------------
ApplyThresholds,3345,2043,940,415,  sum:6.743ms
AveragePool,7,  sum:0.007ms
BatchNorm,201,6,  sum:0.207ms
Convolution,20539,112,  sum:20.651ms
ExtractImagePatches,370,86,11,  sum:0.467ms
QTZ_linear_mid_tread_half,1543,  sum:1.543ms
QuantizedConv2D,316,  sum:0.316ms
QuantizedConv2D_ApplyScalingFactor,6,  sum:0.006ms
ReLu,5,  sum:0.005ms
SoftMax,1,  sum:0.001ms
TotalInitTime,6088,  sum:6.088ms
TotalRunTime,94128,  sum:94.128ms
kn2row,20484,  sum:20.484ms
kn2row-1x1,112,  sum:0.112ms
kn2row-buf,1,  sum:0.001ms
matrix_multiplication,18687,110,  sum:18.797ms
matrix_shift_add1,365,336,178,167,  sum:1.046ms
matrix_shift_add2,1201,526,253,87,  sum:2.067ms
matrix_shift_add_f1,211,  sum:0.211ms
matrix_shift_add_f2,1584,  sum:1.584ms
pack_input_to_qwords,342,668,351,181,91,  sum:1.633ms
quantized-kn2row,31592,24962,8780,5265,314,  sum:70.913ms
quantized_matrix_multiplication,20481,18920,6886,4336,220,  sum:50.843ms
quantized_ohwi_to_hwoi,2,11,3,8,1,  sum:0.025ms
```
after
```
konda@leapmind-server-08:~/work/blueoil/lmnet/saved/konda_20190108151918/export/save.ckpt-957/128x128/output/models/lib$ ./lm_x86.elf input.npy output.npy 
-------------------------------------------------------------
Comparison: Default network test  succeeded!!!
-------------------------------------------------------------
ApplyThresholds,4185,2397,1129,480,  sum:8.191ms
AveragePool,7,  sum:0.007ms
BatchNorm,1148,10,  sum:1.158ms
Convolution,22161,194,  sum:22.355ms
ExtractImagePatches,440,111,74,  sum:0.625ms
QTZ_linear_mid_tread_half,2381,  sum:2.381ms
QuantizedConv2D,452,  sum:0.452ms
QuantizedConv2D_ApplyScalingFactor,5,  sum:0.005ms
ReLu,4,  sum:0.004ms
SoftMax,8,  sum:0.008ms
TotalInitTime,4316,  sum:4.316ms
TotalRunTime,135658,  sum:135.658ms
kn2row,22104,  sum:22.104ms
kn2row-1x1,194,  sum:0.194ms
kn2row-buf,0,  sum:0ms
matrix_multiplication,19531,193,  sum:19.724ms
matrix_shift_add1,376,392,177,168,  sum:1.113ms
matrix_shift_add2,2152,728,525,188,  sum:3.593ms
matrix_shift_add_f1,213,  sum:0.213ms
matrix_shift_add_f2,2355,  sum:2.355ms
pack_input_to_qwords,356,692,346,178,89,  sum:1.661ms
quantized-kn2row,32254,52747,10522,12871,450,  sum:108.844ms
quantized_matrix_multiplication,19452,45987,8185,11780,357,  sum:85.761ms
quantized_ohwi_to_hwoi,1,7,2,8,1,  sum:0.019ms
```
benchmark configuration
```
network
　LmnetV1Quantize
dataset you used
　Caltech101
training configurattion
　epoch is 2. other is default.
hardware
　DE10-nano
how many times you run the benchmark
　As much as average average value is known (around 10 or more)
```