@lm-konda 2019-01-08
AVX2 may contribute not a little.