@iizukak 2019-03-05
Sorry, currently we have no plan to support LLVM.